+++ webpack.dev.js
@@ -1,5 +1,6 @@
 const webpack = require('webpack');
 const merge = require('webpack-merge');
+const path = require('path');
 
 const CommonConfig = require('./webpack.common');
 
@@ -15,11 +16,13 @@ const DevConfig = merge.smartStrategy(
     hot: false,
     historyApiFallback: { disableDotRule: true },
     // host: '0.0.0.0', // Required for Docker -- someone will need to link this somehow
-    publicPath: '/',
-    watchContentBase: true,
     compress: true,
     port: process.env.PORT || 3000,
-    contentBase: 'dist',
+    static: {
+      directory: path.join(__dirname, 'dist'),
+      watch: true,
+      publicPath: '/',
+    }
   },
   module: {
     rules: [